requirement.txt
openmeteo-requests
requests-cache
retry-requests
numpy==1.21.6
pandas==1.3.5
matplotlib==3.5.3
scikit-learn==1.0.2
python-dateutil==2.9.0.post0